+++ wpull/app_test.py
@@ -187,6 +187,7 @@ class TestApp(GoodAppTestCase):
             '--http-compression',
             '--bind-address', '127.0.0.1',
             '--html-parser', 'html5lib',
+            '--link-extractors', 'html',
         ])
         with cd_tempdir():
             builder = Builder(args)